Split `select.rs` into `select.rs`, `select_list.rs`, and `select_dict.rs`.
